long conversion for loss

+++ train_reformulator.py
@@ -108,7 +108,7 @@ def train(args, knn_index, ranking_model, reformulator, optimizer, loss_fn, trai
 
             # new_queries should match document representation of relevant document
             target_embeddings = get_relevant_embeddings(query_id, qrels, knn_index).to(device)
-            batch_loss = loss_fn(new_queries.long(), target_embeddings.long())
+            batch_loss = loss_fn(new_queries, target_embeddings.long())
 
             optimizer.zero_grad()
             batch_loss.backward()
